Move case search models to case_search.py
